 \name{predbayescor-internal}
\alias{abs_cor}
\alias{ find.thresholds}
\alias{ gen.alpha.IGamma}
\alias{ present}
\alias{ present.pair.tex}
\alias{ present.sgl.tex}
\alias{ prpr}
\alias{ real2bin}
\alias{ selfth.abscor}

\title{Internal Functions}

\description{Internal Functions. Type function name directly to see the definition of this function.}

\seealso{
\code{\link{Predict.bayes}}
}

\keyword{internal}